refactor:  compare-state as a component
