+++ src/pages/history/components/comparison/components/EmptyContent/EmptyContent.jsx
