+++ src/pages/history/components/comparison/components/Tree/components/ComplexTree/index.jsx
@@ -1,9 +1,9 @@
 import React, { useState } from "react";
 import PropTypes from "prop-types";
 
-import { Tree } from "pages/compare-json/components/Tree";
+import { Tree } from "pages/history/components/comparison/components/Tree";
 
-import { getIndent, isArray } from "pages/compare-json/utils";
+import { getIndent, isArray } from "pages/history/components/comparison/utils";
 
 export const ComplexTree = (props) => {
   let {
