+++ src/pages/history/components/comparison/styles.js
