+++ src/pages/history/components/comparison/utils/getType.js
