+++ src/pages/history/components/comparison/utils/index.js
