+++ src/pages/history/components/comparison/utils/isArray.js
