+++ src/pages/history/components/comparison/utils/isObject.js
