+++ src/pages/history/components/comparison/utils/needFormat.js
