+++ src/routes/index.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Route, Routes } from "react-router-dom";
 import { Dashboard } from "pages/dashboard";
 import { SignIn } from "pages/sign/sign-in";
 import { Workflows } from "pages/workflows";
-import { CompareJson } from "pages/compare-json";
+// import { CompareJson } from "pages/compare-json";
 import { Processes } from "pages/processes";
 import { History } from "pages/history";
 import { DiagramRefactored } from "pages/diagram";
@@ -25,7 +25,7 @@ export const AppRoutes = () => {
             element={<DiagramRefactored />}
           />
 
-          <Route path="compare-json" element={<CompareJson />} />
+          {/* <Route path="compare-json" element={<CompareJson />} /> */}
         </Route>
       </Routes>
     </BrowserRouter>
